+++ config/ember-try.js
@@ -1,25 +1,27 @@
-/*jshint node:true*/
+/* eslint-env node */
 module.exports = {
   scenarios: [
-    {
-      name: 'ember-lts-2.4',
-      bower: {
-        dependencies: {
-          'ember': 'lts-2-4'
-        },
-        resolutions: {
-          'ember': 'lts-2-4'
-        }
-      }
-    },
     {
       name: 'ember-lts-2.8',
       bower: {
         dependencies: {
-          'ember': 'lts-2-8'
+          'ember': 'components/ember#lts-2-8'
         },
         resolutions: {
           'ember': 'lts-2-8'
+        }
+      },
+      npm: {
+        devDependencies: {
+          'ember-source': null
+        }
+      }
+    },
+    {
+      name: 'ember-lts-2.12',
+      npm: {
+        devDependencies: {
+          'ember-source': '~2.12.0'
         }
       }
     },
@@ -27,10 +29,15 @@ module.exports = {
       name: 'ember-release',
       bower: {
         dependencies: {
-          'ember': 'release'
+          'ember': 'components/ember#release'
         },
         resolutions: {
           'ember': 'release'
+        }
+      },
+      npm: {
+        devDependencies: {
+          'ember-source': null
         }
       }
     },
@@ -38,10 +45,15 @@ module.exports = {
       name: 'ember-beta',
       bower: {
         dependencies: {
-          'ember': 'beta'
+          'ember': 'components/ember#beta'
         },
         resolutions: {
           'ember': 'beta'
+        }
+      },
+      npm: {
+        devDependencies: {
+          'ember-source': null
         }
       }
     },
@@ -49,11 +61,22 @@ module.exports = {
       name: 'ember-canary',
       bower: {
         dependencies: {
-          'ember': 'canary'
+          'ember': 'components/ember#canary'
         },
         resolutions: {
           'ember': 'canary'
         }
+      },
+      npm: {
+        devDependencies: {
+          'ember-source': null
+        }
+      }
+    },
+    {
+      name: 'ember-default',
+      npm: {
+        devDependencies: {}
       }
     }
   ]