+++ tests/helpers/start-app.js
@@ -5,18 +5,13 @@ import config from '../../config/environment';
 const { merge, run } = Ember;
 
 export default function startApp(attrs) {
-  let application;
-
-  // use defaults, but you can override
-  // Ember.assign is not supported on Ember < 2.5
   let attributes = merge({}, config.APP);
-  merge(attributes, attrs);
+  attributes = merge(attributes, attrs); // use defaults, but you can override;
 
-  run(() => {
-    application = Application.create(attributes);
+  return run(() => {
+    let application = Application.create(attributes);
     application.setupForTesting();
     application.injectTestHelpers();
+    return application;
   });
-
-  return application;
 }